[FLINK-21348] Add tests for AdaptiveScheduler's StateWithExecutionGraph
This closes #15342
